move to_json impl to controller
