+++ src/models/mod.rs
@@ -1,5 +1,3 @@
-use std::collections::BTreeMap;
-use rustc_serialize::json::{ToJson, Json};
 use chrono::*;
 #[derive(Default, Debug)]
 pub struct Account {
@@ -18,19 +16,3 @@ pub struct Task {
     pub status          : i32,//0:new,1:ongoing,2:finished,3:canceld
 }
 
-impl ToJson for Task {
-    fn to_json(&self) -> Json {
-        let mut m: BTreeMap<String, Json> = BTreeMap::new();
-        m.insert("id".to_string(), self.id.to_json());
-        m.insert("name".to_string(), self.name.to_json());
-        m.insert("content".to_string(), self.content.to_json());
-        if let Some(dt)=self.create_time{
-            m.insert("create_time".to_string(),dt.format("%Y-%m-%d %H:%M:%S").to_string().to_json());
-        }
-        if let Some(dt)=self.update_time{
-            m.insert("update_time".to_string(),dt.format("%Y-%m-%d %H:%M:%S").to_string().to_json());
-        }
-        m.insert("status".to_string(), self.status.to_json());
-        m.to_json()
-    }
-}
